Bringing initial ros and arduino work into github The ros environment is in "tractor/ros2_ws", and has a partially-tested "TractorWriter" node written (serial echoes out from the arduino properly, but has not been tested to actually control the tractor. Contact Han for more details), a clone of the hokuyo ros2 driver node, desynced from that repository, and a "my_node" that can be repurposed as needed.
On the arduino side, a "passthrough" folder is within, containing a script
for the arduino that should listen to the controller and serial
appropriately (RC control has been tested, echoing serial back to the ros
node has been tested but the whole system together has not been).

Again, contact Han for details/more info.